Detect the point in the frame at which player position is updated
This is almost always at the very end of the frame, indicating that the
position value for the current frame is actually being updated before we
read the playtime value.

However, this change includes some additional logic to detect when this
is not the case, and we can see that this is exceedingly rare, but that
when it does happen, it's within 25% of the frame time. So we could
probably watch for a change in playtime, then wait 4ms of real time,
then read position.