Experiment via Python: we only need cwd and 'SteamAppId' to launch exe
Normally, the game must be launched through Steam: if you just run the
executable from Windows, you'll get an error dialog that says "Failed
to initialize Steam."

Launching the game via Steam and checking it with procexp64 shows a
bunch of environment variables that are supplied by Steam. Luckily, we
only need to do a couple of things to ensure that the game can
initialize Steam when we launch it ourselves:

- Set the 'SteamAppId' environment variable: this is enough to get us
  past the error dialog; luckily we don't need to bother with user IDs
  or other Steam-related environment vars

- Ensure that the game is launched with the install directory (i.e. the
  location of the EXE) as its working directory. If we fail to do this,
  the game window launches but immediately crashes

The game also doesn't check to see whether it's a child of steam.exe or
anything like that; these two changes are sufficient.